refactor: Rename `…ws/package/compile/events/api-gateway/lib/restApi.js`
