+++ lib/plugins/aws/package/compile/events/api-gateway/index.js
@@ -6,7 +6,7 @@ const BbPromise = require('bluebird');
 const memoize = require('memoizee');
 
 const validate = require('./lib/validate');
-const compileRestApi = require('./lib/restApi');
+const compileRestApi = require('./lib/rest-api');
 const compileRequestValidators = require('./lib/request-validator');
 const compileApiKeys = require('./lib/api-keys');
 const compileUsagePlan = require('./lib/usagePlan');
