+++ lib/plugins/aws/package/compile/events/api-gateway/lib/rest-api.js
